+++ gokart/__init__.py
@@ -1,8 +1,9 @@
-from gokart.build import add_config, build
+from gokart.build import build
 from gokart.info import make_tree_info, tree_info
 from gokart.pandas_type_config import PandasTypeConfig
 from gokart.parameter import ExplicitBoolParameter, ListTaskInstanceParameter, TaskInstanceParameter
 from gokart.run import run
 from gokart.task import TaskOnKart
 from gokart.testing import test_run
+from gokart.utils import add_config
 from gokart.workspace_management import delete_local_unnecessary_outputs
